Ajustes

+++ Sistema.Secretaria/Sistema.Secretaria.Application/Alunos/Response/ConsultarNotasResponse.cs
@@ -1,7 +1,7 @@
 ﻿namespace Sistema.Secretaria.Application.Alunos.Response;
 public class ConsultarNotasResponse
 {
-    public double nota_p1 { get; set; }
-    public double nota_p2 { get; set; }
-    public double nota_pf { get; set; }
+    public float nota_p1 { get; set; }
+    public float nota_p2 { get; set; }
+    public float nota_pf { get; set; }
 }
